backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
anthropic==0.25.1
pypdf==4.2.0
pdfplumber==0.11.0
python-docx==1.1.2
pandas==2.2.2
sentence-transformers==3.0.1
numpy==1.26.4
python-dotenv==1.0.1
slowapi==0.1.9
aiofiles==23.2.1
python-magic-bin==0.4.14; sys_platform == "win32"
python-magic==0.4.27; sys_platform != "win32"
httpx==0.27.0
scikit-learn==1.4.2
